+++ src/conf.py
@@ -78,7 +78,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = u'Axom'
-copyright = u'2017-2024, Lawrence Livermore National Security, LLNS'
+copyright = u'2017-2025, Lawrence Livermore National Security, LLNS'
 
 # -- Option for numbering figures/tables/etc.-----------------------------------
 # Note: numfig requires Sphinx (1.3+)